Renaming Projectile compilation unit, so to better conform with standards
Soon others will take the same path.
