+++ src/video.cpp
@@ -9,7 +9,7 @@
 #include "Train.h"
 #include "HeroTrain.h"
 #include "VillainTrain.h"
-#include  "projectile.h"
+#include  "Projectile.h"
 
 #include "video.h"
 #include "generator.h"
